make this work for IPv4 servers on Android:
Android uses IPv6 sockets for IPv4 communication with v4-mapped
addresses (basically, it sends a packet to ::ffff:1.2.3.4 and
the (Linux) kernel running in dual stack mode (not doble stack!)
makes an IPv4 packet addressed to 1.2.3.4 from it). Interestingly
modern.

The problem here is that this requires IPPROTO_IP:IP_RECVTOS to
be setsockopt’d on an AF_INET6 socket (and it will then receive
IPPROTO_IP:IP_TOS cmsg(3)s…) which is… interesting. It will need
to be tested more.