[WIP] blittable parameter block structure for P/Invoke
Even if ecnhll_rcv looks ugly on the C♯ side it’s blittable and
we use a Span<byte> slice to access it anyway to get the IPv6
address when needed. We can throw buffer and parblk across the
invoke boundary and both are handled correctly, v4 and v6 also,
and all *without* unsafe, and I believe with pinning due to this
being blittable.